update API to use AI model

+++ backend/api.py
@@ -6,6 +6,8 @@ import os
 from dotenv import load_dotenv
 import requests
 
+from model import createDataset, createModel, getRec
+
 load_dotenv()
 app = FastAPI()
 
@@ -19,10 +21,10 @@ class Movie(BaseModel):
 def index(userToken: str):
     connection = sqlite3.connect("movie.db")
     cursor = connection.cursor()
-    cursor.execute("SELECT title, genres, imdbId, tmdbId, rating FROM movies INNER JOIN links on movies.movieID = links.movieID LEFT OUTER JOIN ratings on movies.movieId = ratings.movieId AND ratings.userToken = ? LIMIT 8", (userToken,))
+    cursor.execute("SELECT movies.movieId, title, genres, imdbId, tmdbId, rating FROM movies INNER JOIN links on movies.movieId = links.movieId LEFT OUTER JOIN ratings on movies.movieId = ratings.movieId AND ratings.userToken = ? LIMIT 8", (userToken,))
     movies = cursor.fetchall()
     connection.close()
-    movies = [{"title": title, "genres": genres, "tmdbId": tmdbId, "picUrl": retrievePic(tmdbId), "rating": rating} for [title, genres, imdbId, tmdbId, rating] in movies]
+    movies = [{"movieId": movieId, "title": title, "genres": genres, "tmdbId": tmdbId, "picUrl": retrievePic(tmdbId), "rating": rating} for [movieId, title, genres, imdbId, tmdbId, rating] in movies]
     return movies
 
 @app.get("/movies/{movTitle}/{userToken}")
@@ -30,11 +32,12 @@ def search(movTitle: str, userToken):
     print(userToken)
     connection = sqlite3.connect("movie.db")
     cursor = connection.cursor()
-    cursor.execute("SELECT title, genres, imdbId, tmdbId, rating FROM movies INNER JOIN links on movies.movieID = links.movieID LEFT OUTER JOIN ratings on movies.movieId = ratings.movieId AND ratings.userToken = ? WHERE title LIKE ? LIMIT 8", (userToken, '%' + movTitle + '%'))
+    cursor.execute("SELECT * FROM ratings WHERE userToken = ?", (userToken,))
+    rates = cursor.fetchall()
+    cursor.execute("SELECT movies.movieId, title, genres, imdbId, tmdbId, rating FROM movies INNER JOIN links on movies.movieID = links.movieID LEFT OUTER JOIN ratings on movies.movieId = ratings.movieId AND ratings.userToken = ? WHERE title LIKE ? LIMIT 8", (userToken, '%' + movTitle + '%',))
     movies = cursor.fetchall()
-    print(movies)
     connection.close()
-    movies = [{"title": title, "genres": genres, "tmdbId": tmdbId, "picUrl": retrievePic(tmdbId), "rating": rating} for [title, genres, imdbId, tmdbId, rating] in movies]
+    movies = [{"movieId": movieId, "title": title, "genres": genres, "tmdbId": tmdbId, "picUrl": retrievePic(tmdbId), "rating": rating} for [movieId, title, genres, imdbId, tmdbId, rating] in movies]
     return movies
 
 @app.post("/signup/{userToken}")
@@ -69,6 +72,10 @@ def rate(userToken: str, movieId: int, rating: float):
     movie = cursor.fetchall()
     if not movie:
         raise HTTPException(status_code=404, detail="movie doesn't exist")
+    cursor.execute("SELECT * FROM users WHERE userToken = ?", (userToken,))
+    user = cursor.fetchall()
+    if not user:
+        raise HTTPException(status_code=404, detail="user doesn't exist")
     cursor.execute("DELETE FROM ratings WHERE userToken = ? AND movieId = ?", (userToken, movieId))
     cursor.execute("INSERT INTO ratings(userToken, movieId,  rating) VALUES(?, ?, ?)", (userToken, movieId, rating))
     cursor.execute("SELECT * FROM ratings")
@@ -92,6 +99,29 @@ def retrievePic(tmdbId: int):
             return None
     except:
         return None
+
+currentModel = None
+
+@app.post("/train/{userToken}")
+def trainModel():
+    df = createDataset()
+    global currentModel
+    currentModel = createModel(df)
+    return True
+
+@app.get("/recommendation/{userToken}")
+def getRecommendation(userToken: str):
+    rec = getRec(currentModel, userToken)
+    recId = rec[1]
+
+    connection = sqlite3.connect("movie.db")
+    cursor = connection.cursor()
+    cursor.execute("SELECT movies.movieId, title, genres, imdbId, tmdbId, rating FROM movies INNER JOIN links on movies.movieId = links.movieId LEFT OUTER JOIN ratings on movies.movieId = ratings.movieId AND ratings.userToken = ? WHERE movies.movieId = ? LIMIT 1", (userToken, recId))
+    movies = cursor.fetchall()
+    connection.close()
+    movies = [{"movieId": movieId, "title": title, "genres": genres, "tmdbId": tmdbId, "picUrl": retrievePic(tmdbId), "rating": rating, "expected": -rec[0]} for [movieId, title, genres, imdbId, tmdbId, rating] in movies]
+    return movies
+
